Extract Cross into its own component
